refactor method-ref fixes to be used in headless mode.
