+++ org.eclipse.jdt.ui/ui/org/eclipse/jdt/internal/ui/text/correction/QuickAssistProcessor.java
@@ -83,7 +83,6 @@ import org.eclipse.jdt.core.dom.Dimension;
 import org.eclipse.jdt.core.dom.DoStatement;
 import org.eclipse.jdt.core.dom.EnhancedForStatement;
 import org.eclipse.jdt.core.dom.Expression;
-import org.eclipse.jdt.core.dom.ExpressionMethodReference;
 import org.eclipse.jdt.core.dom.ExpressionStatement;
 import org.eclipse.jdt.core.dom.FieldAccess;
 import org.eclipse.jdt.core.dom.FieldDeclaration;
@@ -102,7 +101,6 @@ import org.eclipse.jdt.core.dom.MethodDeclaration;
 import org.eclipse.jdt.core.dom.MethodInvocation;
 import org.eclipse.jdt.core.dom.MethodReference;
 import org.eclipse.jdt.core.dom.Modifier;
-import org.eclipse.jdt.core.dom.Modifier.ModifierKeyword;
 import org.eclipse.jdt.core.dom.Name;
 import org.eclipse.jdt.core.dom.NameQualifiedType;
 import org.eclipse.jdt.core.dom.NumberLiteral;
@@ -111,7 +109,6 @@ import org.eclipse.jdt.core.dom.ParenthesizedExpression;
 import org.eclipse.jdt.core.dom.PostfixExpression;
 import org.eclipse.jdt.core.dom.PrimitiveType;
 import org.eclipse.jdt.core.dom.QualifiedName;
-import org.eclipse.jdt.core.dom.ReturnStatement;
 import org.eclipse.jdt.core.dom.SimpleName;
 import org.eclipse.jdt.core.dom.SimpleType;
 import org.eclipse.jdt.core.dom.SingleVariableDeclaration;
@@ -126,7 +123,6 @@ import org.eclipse.jdt.core.dom.ThrowStatement;
 import org.eclipse.jdt.core.dom.TryStatement;
 import org.eclipse.jdt.core.dom.Type;
 import org.eclipse.jdt.core.dom.TypeDeclaration;
-import org.eclipse.jdt.core.dom.TypeParameter;
 import org.eclipse.jdt.core.dom.UnionType;
 import org.eclipse.jdt.core.dom.VariableDeclaration;
 import org.eclipse.jdt.core.dom.VariableDeclarationExpression;
@@ -154,10 +150,12 @@ import org.eclipse.jdt.internal.corext.dom.Selection;
 import org.eclipse.jdt.internal.corext.dom.SelectionAnalyzer;
 import org.eclipse.jdt.internal.corext.dom.TokenScanner;
 import org.eclipse.jdt.internal.corext.fix.AddInferredLambdaParameterTypesFixCore;
+import org.eclipse.jdt.internal.corext.fix.AddMissingMethodDeclarationFixCore;
 import org.eclipse.jdt.internal.corext.fix.AddVarLambdaParameterTypesFixCore;
 import org.eclipse.jdt.internal.corext.fix.ChangeLambdaBodyToBlockFixCore;
 import org.eclipse.jdt.internal.corext.fix.ChangeLambdaBodyToExpressionFixCore;
 import org.eclipse.jdt.internal.corext.fix.CleanUpConstants;
+import org.eclipse.jdt.internal.corext.fix.CompilationUnitRewriteOperationsFixCore;
 import org.eclipse.jdt.internal.corext.fix.ControlStatementsFix;
 import org.eclipse.jdt.internal.corext.fix.ConvertLambdaToMethodReferenceFixCore;
 import org.eclipse.jdt.internal.corext.fix.ConvertLoopFixCore;
@@ -981,252 +979,19 @@ public class QuickAssistProcessor implements IQuickAssistProcessor {
 		return false;
 	}
 
-	private static class ReturnType {
-		public Type type;
-
-		public ITypeBinding binding;
-	}
-
-	private static ReturnType getReturnType(AST ast, ImportRewrite importRewrite, Type variableType) {
-		ReturnType returnType= new ReturnType();
-		if (variableType instanceof ParameterizedType) {
-			variableType= (Type) ((ParameterizedType) variableType).typeArguments().get(0);
-			ITypeBinding returnTypeBinding= variableType.resolveBinding();
-			if (returnTypeBinding != null) {
-				if (returnTypeBinding.isCapture()) {
-					returnType.binding= returnTypeBinding.getErasure();
-					returnType.type= importRewrite.addImport(returnTypeBinding.getErasure(), ast);
-				} else if (returnTypeBinding.isWildcardType()) {
-					returnType.binding= returnTypeBinding.getBound();
-					returnType.type= importRewrite.addImport(returnTypeBinding.getBound(), ast);
-				} else {
-					returnType.type= importRewrite.addImport(returnTypeBinding, ast);
-					returnType.binding= returnTypeBinding;
-				}
-			}
-		}
-		return returnType;
-	}
-
-	private static Block getNewReturnBlock(AST ast, ITypeBinding returnTypeBinding) {
-		Block newBlock= ast.newBlock();
-		if (!"void".equals(returnTypeBinding.getName())) { //$NON-NLS-1$
-			ReturnStatement newReturnStatement= ast.newReturnStatement();
-			String bName= returnTypeBinding.getBinaryName();
-			if ("Z".equals(bName)) { //$NON-NLS-1$
-				newReturnStatement.setExpression(ast.newBooleanLiteral(false));
-			} else if (returnTypeBinding.isPrimitive()) {
-				newReturnStatement.setExpression(ast.newNumberLiteral());
-			} else if ("java.lang.String".equals(bName)) { //$NON-NLS-1$
-				newReturnStatement.setExpression(ast.newStringLiteral());
-			} else {
-				newReturnStatement.setExpression(ast.newNullLiteral());
-			}
-			newBlock.statements().add(newReturnStatement);
-		}
-		return newBlock;
-	}
-
 	private static boolean getAddMethodDeclaration(IInvocationContext context, ASTNode covering, Collection<ICommandAccess> resultingCollections) {
-		CompilationUnit astRoot= context.getASTRoot();
-		ExpressionMethodReference methodReferenceNode= covering instanceof ExpressionMethodReference
-				? (ExpressionMethodReference) covering
-				: ASTNodes.getParent(covering, ExpressionMethodReference.class);
-		if (methodReferenceNode == null) {
-			return false;
-		}
-		boolean addStaticModifier= false;
-		TypeDeclaration typeDeclaration= ASTNodes.getParent(methodReferenceNode, TypeDeclaration.class);
-
-		if (QuickAssistProcessorUtil.isTypeReferenceToInstanceMethod(methodReferenceNode)) {
-			String methodReferenceQualifiedName= ((Name) methodReferenceNode.getExpression()).getFullyQualifiedName();
-			String typeDeclarationName= astRoot.getPackage().getName().getFullyQualifiedName() + '.' + typeDeclaration.getName().getFullyQualifiedName();
-			if (!methodReferenceQualifiedName.equals(typeDeclarationName)
-					&& !methodReferenceQualifiedName.equals(typeDeclaration.getName().getFullyQualifiedName())) {
-				// only propose for references in same class
-				return false;
-			}
-			addStaticModifier= true;
-		}
-
-		AST ast= astRoot.getAST();
-		ASTRewrite rewrite= ASTRewrite.create(ast);
-		ListRewrite listRewrite= rewrite.getListRewrite(typeDeclaration, TypeDeclaration.BODY_DECLARATIONS_PROPERTY);
-
-		String label= Messages.format(CorrectionMessages.AddUnimplementedMethodReferenceOperation_AddMissingMethod_group,
-				new String[] { methodReferenceNode.getName().getIdentifier(), typeDeclaration.getName().getIdentifier() });
-		Image image= JavaPluginImages.get(JavaPluginImages.IMG_CORRECTION_CHANGE);
-		ASTRewriteCorrectionProposal proposal= new ASTRewriteCorrectionProposal(label, context.getCompilationUnit(), rewrite, IProposalRelevance.ADD_INFERRED_LAMBDA_PARAMETER_TYPES, image);
-		// ImportRewrite importRewrite= proposal.createImportRewrite(context.getASTRoot());
-		ImportRewrite importRewrite= StubUtility.createImportRewrite(astRoot, true);
-
-		VariableDeclarationStatement variableDeclarationStatement= ASTNodes.getParent(methodReferenceNode, VariableDeclarationStatement.class);
-		MethodInvocation methodInvocationNode= ASTNodes.getParent(methodReferenceNode, MethodInvocation.class);
-		Assignment variableAssignment= ASTNodes.getParent(methodReferenceNode, Assignment.class);
-
-		if ((variableAssignment != null || variableDeclarationStatement != null) && methodInvocationNode == null) {
-			/*
-			 * variable declaration
-			 */
-			Type type= null;
-			ReturnType returnType= null;
-			if (variableDeclarationStatement != null) {
-				type= variableDeclarationStatement.getType();
-				returnType= getReturnType(ast, importRewrite, type);
-			} else {
-				Expression leftHandSide= variableAssignment.getLeftHandSide();
-				ITypeBinding assignmentTypeBinding= leftHandSide.resolveTypeBinding();
-				if (assignmentTypeBinding == null) {
-					return false;
-				}
-				type= importRewrite.addImport(assignmentTypeBinding, ast);
-				returnType= new ReturnType();
-				returnType.type= type;
-				returnType.binding= assignmentTypeBinding;
-			}
-			if (returnType.binding == null) {
-				return false;
-			}
-			MethodDeclaration newMethodDeclaration= ast.newMethodDeclaration();
-			newMethodDeclaration.setName((SimpleName) rewrite.createCopyTarget(methodReferenceNode.getName()));
-			newMethodDeclaration.modifiers().add(ast.newModifier(ModifierKeyword.PRIVATE_KEYWORD));
-			if (addStaticModifier) {
-				newMethodDeclaration.modifiers().add(ast.newModifier(ModifierKeyword.STATIC_KEYWORD));
-			}
-
-			IMethodBinding functionalInterfaceMethod= variableDeclarationStatement == null
-					? returnType.binding.getFunctionalInterfaceMethod()
-					: variableDeclarationStatement.getType().resolveBinding().getFunctionalInterfaceMethod();
-			if (functionalInterfaceMethod != null) {
-				returnType.type= importRewrite.addImport(functionalInterfaceMethod.getReturnType(), ast);
-				returnType.binding= functionalInterfaceMethod.getReturnType();
-				ITypeBinding[] typeArguments= functionalInterfaceMethod.getParameterTypes();
-				for (int i= 0; i < typeArguments.length; i++) {
-					ITypeBinding iTypeBinding= typeArguments[i];
-					SingleVariableDeclaration newSingleVariableDeclaration= ast.newSingleVariableDeclaration();
-					newSingleVariableDeclaration.setName(ast.newSimpleName(iTypeBinding.getErasure().getName().toLowerCase() + (i + 1)));
-					newSingleVariableDeclaration.setType(importRewrite.addImport(iTypeBinding.getErasure(), ast));
-					newMethodDeclaration.parameters().add(newSingleVariableDeclaration);
-				}
-			}
-			newMethodDeclaration.setReturnType2(returnType.type);
-			Block newBlock= getNewReturnBlock(ast, returnType.binding);
-			newMethodDeclaration.setBody(newBlock);
-			listRewrite.insertLast(newMethodDeclaration, null);
-
-			// add proposal
-			resultingCollections.add(proposal);
+		if (resultingCollections == null) {
 			return true;
 		}
 
-		/*
-		 * method invocation
-		 */
-		IMethodBinding methodBinding= methodInvocationNode == null ? null : methodInvocationNode.resolveMethodBinding();
-		if (methodBinding == null) {
-			return false;
+		CompilationUnitRewriteOperationsFixCore fix= AddMissingMethodDeclarationFixCore.createAddMissingMethodDeclaration(context.getASTRoot(), covering);
+		if (fix != null) {
+			Image image= JavaPluginImages.get(JavaPluginImages.IMG_CORRECTION_CHANGE);
+			FixCorrectionProposal proposal= new FixCorrectionProposal(fix, null, IProposalRelevance.ADD_INFERRED_LAMBDA_PARAMETER_TYPES, image, context);
+			resultingCollections.add(proposal);
+			return true;
 		}
-		List<ASTNode> arguments= methodInvocationNode.arguments();
-		int index= -1;
-		for (int i= 0; i < arguments.size(); i++) {
-			ASTNode node= arguments.get(i);
-			if (node.equals(methodReferenceNode)) {
-				index= i;
-				break;
-			}
-		}
-		ITypeBinding[] parameterTypes= methodBinding.getParameterTypes();
-		ITypeBinding[] typeArguments= methodBinding.getTypeArguments();
-		if (index >= parameterTypes.length) {
-			// node not found
-			return false;
-		}
-		ITypeBinding[] parameterTypesFunctionalInterface= parameterTypes[index].getFunctionalInterfaceMethod().getParameterTypes();
-		ITypeBinding returnTypeBindingFunctionalInterface= parameterTypes[index].getFunctionalInterfaceMethod().getReturnType();
-		MethodDeclaration newMethodDeclaration= ast.newMethodDeclaration();
-		newMethodDeclaration.modifiers().add(ast.newModifier(ModifierKeyword.PRIVATE_KEYWORD));
-		if (addStaticModifier) {
-			newMethodDeclaration.modifiers().add(ast.newModifier(ModifierKeyword.STATIC_KEYWORD));
-		}
-		Type newReturnType= null;
-		if (returnTypeBindingFunctionalInterface.isPrimitive()) {
-			newReturnType= ast.newPrimitiveType(PrimitiveType.toCode(returnTypeBindingFunctionalInterface.getName()));
-		} else {
-			newReturnType= importRewrite.addImport(returnTypeBindingFunctionalInterface, ast);
-			ITypeBinding[] typeParameters= typeDeclaration.resolveBinding().getTypeParameters();
-			bIf: if (returnTypeBindingFunctionalInterface.isTypeVariable() || returnTypeBindingFunctionalInterface.isParameterizedType()) {
-				for (ITypeBinding typeParameter : typeParameters) {
-					// check if parameter type is a Type parameter of the class
-					if (Bindings.equals(typeParameter, returnTypeBindingFunctionalInterface)) {
-						break bIf;
-					}
-				}
-				TypeParameter newTypeParameter= ast.newTypeParameter();
-				newTypeParameter.setName(ast.newSimpleName(returnTypeBindingFunctionalInterface.getName()));
-				addIfMissing(newMethodDeclaration, newTypeParameter);
-			}
-		}
-		newMethodDeclaration.setName((SimpleName) rewrite.createCopyTarget(methodReferenceNode.getName()));
-		newMethodDeclaration.setReturnType2(newReturnType);
-		pLoop: for (int i= 0; i < parameterTypesFunctionalInterface.length; i++) {
-			ITypeBinding parameterType2= parameterTypesFunctionalInterface[i];
-			SingleVariableDeclaration newSingleVariableDeclaration= ast.newSingleVariableDeclaration();
-			if (parameterType2.isCapture()) {
-				newSingleVariableDeclaration.setName(ast.newSimpleName(parameterType2.getErasure().getName().toLowerCase() + (i + 1)));
-				newSingleVariableDeclaration.setType(importRewrite.addImport(parameterType2.getErasure(), ast));
-			} else {
-				newSingleVariableDeclaration.setName(ast.newSimpleName(parameterType2.getName().toLowerCase() + (i + 1)));
-				newSingleVariableDeclaration.setType(importRewrite.addImport(parameterType2, ast));
-			}
-			newMethodDeclaration.parameters().add(newSingleVariableDeclaration);
-			ITypeBinding[] typeParameters= typeDeclaration.resolveBinding().getTypeParameters();
-			if (parameterType2.isTypeVariable()) {
-				// check if parameter type is a Type parameter of the class
-				for (ITypeBinding typeParameter : typeParameters) {
-					if (Bindings.equals(typeParameter, parameterType2)) {
-						continue pLoop;
-					}
-				}
-
-				TypeParameter newTypeParameter= ast.newTypeParameter();
-				newTypeParameter.setName(ast.newSimpleName(importRewrite.addImport(parameterType2)));
-				ITypeBinding[] typeBounds= parameterType2.getTypeBounds();
-				for (ITypeBinding typeBound : typeBounds) {
-					newTypeParameter.typeBounds().add(importRewrite.addImport(typeBound, ast));
-				}
-				addIfMissing(newMethodDeclaration, newTypeParameter);
-			}
-		}
-		for (int i= 0; i < typeArguments.length; i++) {
-			ITypeBinding typeArgument= typeArguments[i];
-			SingleVariableDeclaration newSingleVariableDeclaration= ast.newSingleVariableDeclaration();
-			newSingleVariableDeclaration.setName(ast.newSimpleName(typeArgument.getName().toLowerCase() + (i + 1)));
-			newSingleVariableDeclaration.setType(importRewrite.addImport(typeArgument, ast));
-			newMethodDeclaration.parameters().add(newSingleVariableDeclaration);
-			if (typeArgument.isTypeVariable()) {
-				TypeParameter newTypeParameter= ast.newTypeParameter();
-				newTypeParameter.setName(ast.newSimpleName(importRewrite.addImport(typeArgument)));
-				newMethodDeclaration.typeParameters().add(newTypeParameter);
-			}
-		}
-		Block newBlock= getNewReturnBlock(ast, returnTypeBindingFunctionalInterface);
-		newMethodDeclaration.setBody(newBlock);
-		listRewrite.insertLast(newMethodDeclaration, null);
-
-		// add proposal
-		resultingCollections.add(proposal);
-		return true;
-	}
-
-	private static void addIfMissing(MethodDeclaration methodDeclaration, TypeParameter newTypeParameter) {
-		List<TypeParameter> typeParameters= methodDeclaration.typeParameters();
-		for (TypeParameter typeParameter : typeParameters) {
-			boolean equals= typeParameter.getName().getFullyQualifiedName().equals(newTypeParameter.getName().getFullyQualifiedName());
-			if (equals) {
-				return;
-			}
-		}
-		typeParameters.add(newTypeParameter);
+		return false;
 	}
 
 	public static boolean getInferDiamondArgumentsProposal(IInvocationContext context, ASTNode node, IProblemLocationCore[] locations, Collection<ICommandAccess> resultingCollections) {
